Support schemas with no props
(feature) You can now define schemas with no props. If you also set
additionalProperties to true, a schema like that can act like an ANY type.
You can also nest schemas like that inside your object. This allows defining a
property that can be an object of any shape. You can still add method to a
model like that

(change) Restructured the library code so now all code is in the lib directory
wherease the main index.js just re-exports stuff from within the lib folder.
