+++ lib/util.js
@@ -28,7 +28,44 @@ function mapObject(obj, mapperFn) {
 }
 
 
+// A little helper to manage JSONSchema string formats.
+class FormatManager {
+  constructor() {
+    this.formats = {};
+  }
+
+  // Register new string format with the given name
+  register(name, { dump, load }) {
+    this.formats[name] = { dump, load };
+  }
+
+  // Find string format by name.
+  find(name) {
+    if (!name)
+      return undefined;
+
+    return this.formats[name];
+  }
+
+  loadValue(format, value) {
+    const formatter = this.find(format);
+    return formatter ? formatter.load(value) : value;
+  }
+}
+
+
+// Check if the given object is empty
+function isEmpty(obj) {
+  if (!obj)
+    return true;
+
+  return Object.keys(obj).length === 0;
+}
+
+
 module.exports = {
   json,
   mapObject,
+  isEmpty,
+  FormatManager,
 };
